Modify to include point cloud iteration
read_zdf.py is modified and renamed to read_iterate_zdf.py
