+++ read_iterate_zdf.py
@@ -1,0 +1,50 @@
+"""
+This example shows how to import, display and iterate over a Zivid point cloud from a .ZDF
+file.
+"""
+
+import numpy as np
+import zivid
+
+
+def _main():
+
+    app = zivid.Application()
+
+    # The Zivid3D.zdf file has to be in the same folder as this sample script.
+    filename_zdf = "Zivid3D.zdf"
+
+    print(f"Reading {filename_zdf} point cloud")
+    frame = zivid.Frame(filename_zdf)
+
+    # Extracting point cloud from the frame
+    point_cloud = frame.get_point_cloud().to_array()
+    xyz = np.dstack([point_cloud["x"], point_cloud["y"], point_cloud["z"]])
+    rgb = np.dstack([point_cloud["r"], point_cloud["g"], point_cloud["b"]])
+    contrast = np.dstack([point_cloud["contrast"]])
+
+    height = frame.get_point_cloud().height
+    width = frame.get_point_cloud().width
+
+    print(f"Point cloud information:")
+    print(f"Number of points: {point_cloud.size}")
+    print(f"Height: {height}, Width: {width}")
+
+    # Iterating over the point cloud and displaying X, Y, Z, R, G, B, and Contrast
+    # for central 10 x 10 pixels
+    pixels_to_display = 10
+    for i in range(
+        int((height - pixels_to_display) / 2), int((height + pixels_to_display) / 2)
+    ):
+        for j in range(
+            int((width - pixels_to_display) / 2), int((width + pixels_to_display) / 2)
+        ):
+            print(
+                f"Values at pixel ({i} , {j}): X:{xyz[i,j,0]:.1f} Y:{xyz[i,j,1]:.1f}"
+                f"Z:{xyz[i,j,2]:.1f} R:{rgb[i,j,0]} G:{rgb[i,j,1]} B:{rgb[i,j,2]}"
+                f"Contrast:{contrast[i,j,0]:.1f}"
+            )
+
+
+if __name__ == "__main__":
+    _main()
