--- read_zdf.py
@@ -1,26 +1,0 @@
-"""
-Import ZDF point cloud.
-"""
-
-import numpy as np
-import zivid
-
-
-def _main():
-
-    app = zivid.Application()
-
-    # The Zivid3D.zdf file has to be in the same folder as this sample script.
-    filename_zdf = "Zivid3D.zdf"
-
-    print(f"Reading {filename_zdf} point cloud")
-    frame = zivid.Frame(filename_zdf)
-
-    # Getting the point cloud
-    point_cloud = frame.get_point_cloud().to_array()
-    xyz = np.dstack([point_cloud["x"], point_cloud["y"], point_cloud["z"]])
-    rgb = np.dstack([point_cloud["r"], point_cloud["g"], point_cloud["b"]])
-
-
-if __name__ == "__main__":
-    _main()
